# server/Api/openapi-with-docs.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
    "/codes": {
      "post": {
        "tags": [
          "Action"
        ],
        "operationId": "Action_AddCode",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryCodeRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryCodeDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Action"
        ],
        "operationId": "Action_GetCodes",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntryCodeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/codes/{id}": {
      "delete": {
        "tags": [
          "Action"
        ],
        "operationId": "Action_DeleteCode",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EntryCodeDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "code": {
            "type": "string"
          },
          "codeOwner": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          },
          "type": {
            "$ref": "#/components/schemas/EntryCodeTypeDto"
          },
          "expiry": {
            "type": "string",
            "format": "date-time"
          },
          "useCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "userName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "EntryCodeTypeDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "maxUses": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "CreateEntryCodeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "typeId",
          "expiry"
        ],
        "properties": {
          "typeId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "codeOwnerId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "expiry": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          }
        }
      },
...
    }
  }
}